major refactor
Rename the low-level calls for getting a lock `acquire-*-lock', as in
Bordeaux Threads.  They return true when they get the lock, and false
when there is a timeout that expires, in the spirit of Bordeaux Threads.

Don't use timeouts provided by the system.  They don't work.  At least
not in SBCL.  Instead, loop calls until some test passes or the timeout
expires.

  (See `do-until-timeout' for implementation details.  Basically,
   repeatedly call a thunk until some side effect happens.  For this to
   work, the thunk must return quickly.)

Signal an error when a thread attempts to get the access lock when it
already owns a read-only lock.  This is a program error (you should have
requested an access lock in the first place).

  (The other way is fine, and is required by the internal mechanics of
   the library.)

Change all keyword arguments to optional positional arguments.  It makes
no sense to have a timeout callback function without a timeout.