configurable slot access timeout
Slot reading and writing may now be timed out instead of deadlocking,
depending on the values of *SLOT-READ-TIMEOUT* and *SLOT-WRITE-TIMEOUT*.
The default value is NIL, to not time out.

On timeout, the serious condition LOCK-TIMEOUT is signalled.  Program
authors may use the LOCK-TIMEOUT function to signal this condition in
their own lock timeout callbacks.